Implementacao de listagem de tarefas correspondentes a determinado usuario e respectivas acoes (edit, create, remove)

+++ ListaDeTarefas/gerenciador_de_tarefas/app/views/tarefa_views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from ..forms import TarefaForm
 from ..entidades.tarefa import Tarefa
@@ -7,9 +8,8 @@ from ..services import tarefa_service
 
 @login_required()
 def listar_tarefas(request):
-    tarefas = tarefa_service.listar_tarefas()
+    tarefas = tarefa_service.listar_tarefas(request.user)
     # capturando todas as tarefas
-
     return render(request, 'tarefas/listar_tarefas.html', {"tarefas": tarefas})
 
 @login_required()
@@ -24,7 +24,8 @@ def cadastrar_tarefa(request):
            tarefa_nova = Tarefa(tituloo=tituloo,
                                  descricao=descricao,
                                  data_expiracao=data_expiracao,
-                                 prioridade=prioridade)
+                                 prioridade=prioridade,
+                                 usuario=request.user)
            tarefa_service.cadastrar_tarefa(tarefa_nova)
            return redirect('listar_tarefas')
     else:
@@ -35,6 +36,8 @@ def cadastrar_tarefa(request):
 def editar_tarefa(request, id):
     # buscar qual tarefa quero editar no bd
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não permitido")
     # criamos o form com base nas informacoes que vieram do banco de dados
     form_tarefa = TarefaForm(request.POST or None, instance=tarefa_bd)
     # quando alterarmos as infos e clicar em 'salvar' verificamos se as mesmas sao validas
@@ -48,7 +51,8 @@ def editar_tarefa(request, id):
         tarefa_nova = Tarefa(tituloo=tituloo,
                              descricao=descricao,
                              data_expiracao=data_expiracao,
-                             prioridade=prioridade)
+                             prioridade=prioridade,
+                             usuario=request.user)
         # Enviamos tanto a tarefa antiga(bd) quanto a tarefa nova pro metod editar_tarefa
         tarefa_service.editar_tarefa(tarefa_bd, tarefa_nova)
         return redirect('listar_tarefas')
@@ -58,6 +62,8 @@ def editar_tarefa(request, id):
 def remover_tarefa(request, id):
     # retorna o id da tarefa a ser removida
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não permitido")
     if request.method == "POST":
         tarefa_service.remover_tarefa(tarefa_bd)
         return redirect('listar_tarefas')
